apps/backend/requirements.txt
pgeocode
iso8601
countrycode
haversine
geopy
amadeus
geonamescache
django
django-request
django-cors-headers
python-dotenv

pandas
numpy
anyio
distro
httpx
asgiref
pyarrow
routing~=0.2.0
zstandard
cache
openai
pydantic
elastic
polars
h3

